Add notification-queue
This script runs on desktop login by Freedesktop Autostart, and
sends a notification via notify-send for each line a the FIFO file that
is constantly monitored.

This way, the backup (or any other process) that runs for example with
systemd (by root) can notify the user with the latest backup stats.